How to Hack Hacker News (Project)
Hacker News is a popular website run by Y Combinator. It’s widely known by people in the tech industry as a community site for sharing news, showing off projects, asking questions, among other things.

In this project, you will be working with a table named hacker_news that contains stories from Hacker News since its launch in 2007. It has the following columns:

title - the title of the story
user - the user who submitted the story
score - the score of the story
timestamp - the time of the story
url - the link of the story